Tried adding parenthesis check.
This might be better in the Reader rather than the tokenizer, but that
means we lose all the line and column information in the current
implementation. I think that things are simpler if we have a Token
object which can hold info about where in the source we got the token
for error reporting.