Move some files around.
Old tokenizer class is still there, most reader tests fail. The idea is
to have the reader consume one token ahead in next() and save the result
to some buffer in the Reader class. The caller then invokes peek() to
read the current token.

Note to self - this is starting to sound more and more indirect...